+++ .env.example
@@ -5,8 +5,6 @@ TEST_CFA_KEEP_DB=no
 
 # sqlite | postgresql
 TEST_CFA_FIX_DB=postgresql
-# prisma | none
-TEST_CFA_FIX_ORM=prisma
 
 # Test Postgres server
 TEST_PG_PORT=5432
