+++ .github/workflows/nodejs.yml
@@ -44,12 +44,10 @@ jobs:
         env:
           TEST_CFA_RANDOM_NUM: 1
           TEST_CFA_FIX_DB: postgresql
-          TEST_CFA_FIX_ORM: prisma
       - run: npm test
         env:
           TEST_CFA_RANDOM_NUM: 1
           TEST_CFA_FIX_DB: sqlite
-          TEST_CFA_FIX_ORM: prisma
       - run: npm test
         if: github.event_name == 'schedule'
         env:
