+++ server/api/dbConnection/controller.ts
@@ -11,9 +11,7 @@ export default defineController(() => ({
       return { status: 200, body: { enabled: true } }
     }
 
-    if (answers.orm === 'none' || answers.db === 'sqlite') {
-      return { status: 200, body: { enabled: true } }
-    }
+    if (answers.db === 'sqlite') return { status: 200, body: { enabled: true } }
 
     const templateCtx = answersToTemplateContext({ ...answers, serverPort: 0, clientPort: 0 })
