+++ server/templates/.vscode/extensions.json
@@ -1,3 +1,3 @@
 {
-  "recommendations": ["dbaeumer.vscode-eslint<% if (orm === 'prisma') { %>", "prisma.prisma<% } %>"]
+  "recommendations": ["dbaeumer.vscode-eslint", "prisma.prisma"]
 }
