--- server/templates/@orm=none/@server-dev-dep
@@ -1,3 +1,0 @@
-esbuild
-esbuild-node-externals
-node-dev
