--- server/templates/@orm=prisma/@dep
@@ -1,2 +1,0 @@
-@prisma/client
-prisma
