--- server/templates/@orm=prisma/@server-dep
@@ -1,1 +1,0 @@
-@prisma/client
