--- server/templates/@orm=prisma/@server-dev-dep
@@ -1,4 +1,0 @@
-prisma
-esbuild
-esbuild-node-externals
-node-dev
