+++ server/templates/package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "<% if (orm === 'prisma') { %>npm run migrate:dev && <% } %>run-p dev:*",
+    "dev": "npm run migrate:dev && run-p dev:*",
     "dev:client": "next dev -p <%= clientPort %>",
     "dev:server": "npm run dev --prefix server",
     "dev:aspida": "aspida --watch",
@@ -17,11 +17,11 @@
     "generate:pathpida": "pathpida --enableStatic --ignorePath .gitignore",
     "generate:server": "npm run generate --prefix server",
     "lint": "eslint --ext .ts,.js,.tsx . && prettier --check \"./**/*.{ts,js,tsx}\"",
-    "lint:fix": "eslint --ext .ts,.js,.tsx . --fix && prettier --write \"./**/*.{ts,js,tsx}\"<% if (orm === 'prisma') { %>",
+    "lint:fix": "eslint --ext .ts,.js,.tsx . --fix && prettier --write \"./**/*.{ts,js,tsx}\"",
     "migrate:dev": "npm run migrate:dev --prefix server",
     "migrate:dev:createonly": "npm run migrate:dev:createonly --prefix server",
-    "migrate:reset": "npm run migrate:reset --prefix server<% } %><% if (orm === 'prisma') { %>",
-    "seed": "npm run seed --prefix server<% } %>",
+    "migrate:reset": "npm run migrate:reset --prefix server",
+    "seed": "npm run seed --prefix server",
     "start": "run-p start:*",
     "start:client": "next start",
     "start:server": "npm start --prefix server<% if (testing === 'jest') { %>",
