--- server/templates/server/@orm=prisma/scripts/build.dev.js
@@ -1,4 +1,0 @@
-const { build } = require('esbuild')
-const config = require('./config.dev.js')
-
-build(config)
