--- server/templates/server/@orm=prisma/scripts/build.prod.js
@@ -1,4 +1,0 @@
-const { build } = require('esbuild')
-const config = require('./config.prod.js')
-
-build(config).catch(() => process.exit(1))
