--- server/templates/server/@orm=prisma/scripts/config.dev.js
@@ -1,8 +1,0 @@
-const config = require('./config.common')
-
-module.exports = Object.assign(config, {
-  watch: true,
-  define: {
-    'process.env.NODE_ENV': `"development"`
-  }
-})
