--- server/templates/server/@orm=prisma/scripts/config.prod.js
@@ -1,8 +1,0 @@
-const config = require('./config.common')
-
-module.exports = Object.assign(config, {
-  minify: true,
-  define: {
-    'process.env.NODE_ENV': `"production"`
-  }
-})
