--- server/templates/server/@orm=prisma/types/index.ts
@@ -1,9 +1,0 @@
-export type UserInfo = {
-  id: string
-  name: string
-  icon: string
-}
-
-export type AuthHeader = {
-  authorization: string
-}
