+++ server/templates/server/@server-dep
@@ -1,2 +1,3 @@
+@prisma/client
 dotenv
 zod
