+++ server/templates/server/@server=express/entrypoints/index.ts
