+++ server/templates/server/@server=fastify/entrypoints/index.ts
