+++ server/templates/server/@server=fastify/service/app.ts
