+++ server/templates/server/package.json
@@ -3,20 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "npm run generate && <% if (orm === 'prisma') { %>npm run migrate:dev && <% } %>run-p dev:*",
+    "dev": "npm run generate && npm run migrate:dev && run-p dev:*",
     "dev:server": "node-dev --notify=false --respawn index.js",
     "dev:build": "node ./scripts/build.dev.js",
-    "dev:frourio": "frourio<% if (server === 'express') { %>-express<% } %> --watch<% if (orm === 'prisma') { %>",
-    "dev:prisma": "prisma generate --watch<% } %>",
+    "dev:frourio": "frourio<% if (server === 'express') { %>-express<% } %> --watch",
+    "dev:prisma": "prisma generate --watch",
     "build": "npm run generate && node ./scripts/build.prod.js",
     "generate": "run-p generate:*",
-    "generate:frourio": "frourio<% if (server === 'express') { %>-express<% } %><% if (orm === 'prisma') { %>",
-    "generate:prisma": "prisma generate<% } %><% if (orm === 'prisma') { %>",
+    "generate:frourio": "frourio<% if (server === 'express') { %>-express<% } %>",
+    "generate:prisma": "prisma generate",
     "migrate:dev": "prisma migrate dev",
     "migrate:dev:createonly": "prisma migrate dev --create-only",
-    "migrate:reset": "prisma migrate reset<% } %><% if (orm === 'prisma'){ %>",
+    "migrate:reset": "prisma migrate reset",
     "seed": "prisma db seed --preview-feature",
-    "ts-node": "ts-node --compiler-options \"{\\\"module\\\":\\\"commonjs\\\"}\"<% } %>",
+    "ts-node": "ts-node --compiler-options \"{\\\"module\\\":\\\"commonjs\\\"}\"",
     "start": "cross-env NODE_ENV=production node index.js"
   }
 }
