+++ server/templates/server/prisma/.env
