+++ server/templates/server/prisma/.env.example
