+++ server/templates/server/prisma/schema.prisma
