+++ server/templates/server/prisma/seed.ts
