+++ server/templates/server/scripts/build.dev.js
