+++ server/templates/server/scripts/build.prod.js
