+++ server/templates/server/scripts/config.common.js
