+++ server/templates/server/scripts/config.dev.js
