+++ server/templates/server/scripts/config.prod.js
