+++ server/templates/server/service/tasks.ts
@@ -1,59 +1,22 @@
-import fs from 'fs'<% if (testing !== 'none') { %>
-import { depend } from 'velona'<% } %>
-import type { Task } from '$/types'
+<% if (testing !== 'none') { %>import { depend } from 'velona'
+<% } %>import { PrismaClient, Prisma } from '@prisma/client'
 
-type DB = {
-  nextId: number
-  tasks: Task[]
-}
-
-let dbPath = ''
-
-const readDB = async (): Promise<DB> =>
-  JSON.parse(await fs.promises.readFile(dbPath, 'utf8'))
-const writeDB = (db: DB) =>
-  fs.promises.writeFile(dbPath, JSON.stringify(db), 'utf8')
-
-export const createDBFileIfNotExists = (dbFilePath: string) => {
-  dbPath = dbFilePath
-
-  if (fs.existsSync(dbPath)) return
-
-  fs.writeFileSync(dbPath, JSON.stringify({ nextId: 0, tasks: [] }), 'utf8')
-}
+const prisma = new PrismaClient()
 
 export const getTasks = <% if (testing === 'none') { %>async (limit?: number) =>
-  (await readDB()).tasks.slice(0, limit)<% } else { %>depend({ readDB }, async ({ readDB }, limit?: number) =>
-  (await readDB()).tasks.slice(0, limit)
+  (await prisma.task.findMany()).slice(0, limit)<% } else { %>depend(
+  { prisma: prisma as { task: { findMany(): Promise<Task[]> } } },
+  async ({ prisma }, limit?: number) =>
+    (await prisma.task.findMany()).slice(0, limit)
 )<% } %>
 
-export const createTask = async (label: Task['label']) => {
-  const db = await readDB()
-  const task = { id: db.nextId, label, done: false }
-  await writeDB({ nextId: db.nextId + 1, tasks: [...db.tasks, task] })
-  return task
-}
+export const createTask = (label: Task['label']) =>
+  prisma.task.create({ data: { label } })
 
-export const updateTask = async (
+export const updateTask = (
   id: Task['id'],
-  partialTask: Partial<Pick<Task, 'label' | 'done'>>
-) => {
-  const db = await readDB()
-  const task = db.tasks.find((t) => t.id === id)
-  if (!task) return
+  partialTask: Prisma.TaskUpdateInput
+) => prisma.task.update({ where: { id }, data: partialTask })
 
-  task.label = partialTask.label ?? task.label
-  task.done = partialTask.done ?? task.done
-  await writeDB({
-    nextId: db.nextId,
-    tasks: db.tasks.map((t) => (t.id === id ? task : t))
-  })
-}
-
-export const deleteTask = async (id: Task['id']) => {
-  const db = await readDB()
-  await writeDB({
-    nextId: db.nextId,
-    tasks: db.tasks.filter((t) => t.id !== id)
-  })
-}
+export const deleteTask = (id: Task['id']) =>
+  prisma.task.delete({ where: { id } })
