+++ server/templates/tsconfig.json
@@ -5,8 +5,7 @@
     "baseUrl": ".",
     "paths": {
       "~/*": ["./src/*"],
-      "$/*": ["./server/*"]<% if (orm === 'prisma') { %>,
-      "$prisma/*": ["./server/node_modules/.prisma/*"]<% } %>
+      "$/*": ["./server/*"]
     },
     "allowJs": true,
     "skipLibCheck": true,
