+++ server/tsconfig.json
@@ -1,9 +1,7 @@
 {
   "compilerOptions": {
     "downlevelIteration": true,
-    "experimentalDecorators": true,
     "skipLibCheck": true,
-    "emitDecoratorMetadata": true,
     "esModuleInterop": true,
     "resolveJsonModule": true,
     "lib": ["ES2019"],
