+++ tsconfig.json
@@ -1,26 +1,16 @@
 {
   "compilerOptions": {
     "target": "es5",
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "ES2019"
-    ],
+    "lib": ["dom", "dom.iterable", "ES2019"],
     "baseUrl": ".",
     "paths": {
-      "~/*": [
-        "./*"
-      ],
-      "$/*": [
-        "./server/*"
-      ]
+      "~/*": ["./*"],
+      "$/*": ["./server/*"]
     },
     "allowJs": true,
     "skipLibCheck": true,
     "strict": true,
     "strictPropertyInitialization": false,
-    "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
     "forceConsistentCasingInFileNames": true,
     "noEmit": true,
     "esModuleInterop": true,
@@ -31,13 +21,6 @@
     "jsx": "preserve",
     "incremental": true
   },
-  "exclude": [
-    "node_modules",
-    "server"
-  ],
-  "include": [
-    "next-env.d.ts",
-    "**/*.ts",
-    "**/*.tsx"
-  ]
+  "exclude": ["node_modules", "server"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"]
 }
